답변 수정

+++ server/src/main/java/com/preproject/server/answer/controller/AnswerController.java
@@ -21,7 +21,7 @@ import java.util.List;
  */
 
 @RestController
-@RequestMapping  // url 개별 지정
+@RequestMapping("/question/{question-id}/answer")
 @Validated
 public class AnswerController {
 
@@ -36,7 +36,7 @@ public class AnswerController {
     }
 
     // 답변 등록
-    @PostMapping("/question/{question-id}/answer")
+    @PostMapping()
     public ResponseEntity postAnswer(@Positive @PathVariable("question-id") long questionId,
                                      @Valid @RequestBody AnswerPostDto answerPostDto) {
 
@@ -52,12 +52,20 @@ public class AnswerController {
     }
 
     // 답변 수정 (답변 id를 통해서 조회 후 수정)
-    @PatchMapping("/answer/{answer-id}")
-    public ResponseEntity patchAnswer(@PathVariable("answer-id") long answerId,
+    @PatchMapping("/{answer-id}")
+    public ResponseEntity patchAnswer(@PathVariable("question-id") long questionId,
+                                      @PathVariable("answer-id") long answerId,
                                       @Valid @RequestBody AnswerPatchDto answerPatchDto) {
 
         // 수정할 답변 찾기
         Answer updateAnswer = answerService.findAnswerById(answerId);
+
+
+        // 답변할 질문 찾기
+        if (updateAnswer.getQuestion().getQuestionId() != questionId) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("답변하려는 질문이 없습니다.");
+        }
 
 
         // 답변 내용 업데이트 할 때 null 값인지 여부 확인
@@ -74,12 +82,22 @@ public class AnswerController {
     }
 
     // 답변 조회 ( 질문과 함께 조회가 되어야 함 )
-    // 답변이 등록된걸로 고유의 답변id가 부여됐을테니 질문번호를 찾지 않고 로직 구현
+    // 답변이 등록된걸로 고유의 답변id가 부여됐을테니 질문번호를 찾지 않고 로직 구현 -> 질문 번호 별 답변을 찾는 걸로 수정
     // 답변이 한 페이지에 여러 개의 답변이 구현되려면 페이지네이션을 적용하면 될듯함!
-    @GetMapping("/answer/{answer-id}")
-    public ResponseEntity getAnswer(@PathVariable("answer-id") long answerId) {
+    @GetMapping("/{answer-id}")
+    public ResponseEntity getAnswer(@PathVariable("question-id") long questionId,
+                                    @PathVariable("answer-id") long answerId) {
 
-        // 조회할 질문은 PathVariable의 경로를 통해 해당 값으로 조회 됨
+        // 조회할 답변 찾기
+        Answer findAnswer = answerService.findAnswerById(answerId);
+
+
+        // 답변할 질문 찾기
+        if (findAnswer.getQuestion().getQuestionId() != questionId) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("조회하려는 질문이 없습니다.");
+        }
+
 
         Answer response = answerService.findAnswer(answerId);
 
@@ -89,7 +107,7 @@ public class AnswerController {
     }
 
     // 페이지네이션을 적용하여 조회(질문 번호당)
-    @GetMapping("/question/{question-id}/answer")
+    @GetMapping()
     public ResponseEntity getAnswers(@PathVariable("question-id") long questionId,
                                      @Positive @RequestParam int page,
                                      @Positive @RequestParam int size) {
@@ -107,11 +125,19 @@ public class AnswerController {
     }
 
 
-    // 답변 삭제 (답변 id를 통해서 조회 후 삭제)
-    @DeleteMapping("/answer/{answer-id}")
-    public ResponseEntity deleteAnswer(@PathVariable("answer-id") long answerId) {
+    // 답변 삭제 (질문id에서 먼저 조회 후 답변 id를 통해서 삭제)
+    @DeleteMapping("/{answer-id}")
+    public ResponseEntity deleteAnswer(@PathVariable("question-id") long questionId,
+                                       @PathVariable("answer-id") long answerId) {
 
-        // 삭제할 답변은 PathVarible의 경로를 통해 해당 값으로 조회 됨
+        // 삭제할 답변 조회
+        Answer deleteAnswer1 = answerService.findAnswer(answerId);
+
+        // 삭제할 질문 찾기
+        if(deleteAnswer1.getQuestion().getQuestionId() != questionId) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("삭제할 답변이 없습니다.");
+        }
 
         // 답변 삭제
         answerService.deleteAnswer(answerId);
